[MIST-1068] Move queryStarter and queryRemover from Group to ApplicationInfo
