+++ mist-core/src/main/java/edu/snu/mist/core/task/groupaware/Group.java
@@ -17,8 +17,6 @@ package edu.snu.mist.core.task.groupaware;
 
 import edu.snu.mist.core.task.ExecutionDags;
 import edu.snu.mist.core.task.Query;
-import edu.snu.mist.core.task.QueryRemover;
-import edu.snu.mist.core.task.QueryStarter;
 import edu.snu.mist.core.task.groupaware.eventprocessor.EventProcessor;
 import edu.snu.mist.formats.avro.GroupCheckpoint;
 import org.apache.reef.tang.annotations.DefaultImplementation;
@@ -153,16 +151,6 @@ public interface Group extends AutoCloseable {
    * The number of active queries.
    */
   int size();
-
-  /**
-   * Get the query starter for this group.
-   */
-  QueryStarter getQueryStarter();
-
-  /**
-   * Get the query remover for this group.
-   */
-  QueryRemover getQueryRemover();
 
   /**
    * Get the ExecutionDags of this group.
